refactor: Atualizando graficos de aspectos sociais e desacoplando metodos
feat: Novo grafico de distribuição por estado
