+++ Dashboard.py
@@ -1,10 +1,10 @@
 import streamlit as st
 # Importar módulos do projeto
-from utils.data_loader import load_data, filter_data_by_states
 from utils.mappings import get_mappings
 from tabs.geral import render_geral
 from tabs.aspectos_sociais import render_aspectos_sociais
 from tabs.desempenho import render_desempenho
+from utils.data_loader import load_data, filter_data_by_states
 
 
 # Configuração inicial da página
