+++ dtypes - Copia.json
@@ -1,0 +1,26 @@
+{
+  "TP_FAIXA_ETARIA": "category",
+  "TP_SEXO": "category",
+  "TP_ESTADO_CIVIL": "category",
+  "TP_COR_RACA": "category",
+  "TP_DEPENDENCIA_ADM_ESC": "category",
+  "TP_ST_CONCLUSAO": "category",
+  "SG_UF_PROVA": "category",
+  "TP_PRESENCA_CN": "category",
+  "TP_PRESENCA_CH": "category",
+  "TP_PRESENCA_LC": "category",
+  "TP_PRESENCA_MT": "category",
+  "NU_NOTA_CN": "float16",
+  "NU_NOTA_CH": "float16",
+  "NU_NOTA_LC": "float16",
+  "NU_NOTA_MT": "float16",
+  "NU_NOTA_REDACAO": "int16",
+  "TP_PRESENCA_GERAL": "category",
+  "TP_ANO_CONCLUIU": "category",
+  "NU_DESEMPENHO": "category",
+  "Q001": "category",
+  "Q002": "category",
+  "Q005": "category",
+  "Q006": "category",
+  "Q025": "category"
+}
